+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2018-08-14
+### Fixed
+- Fix issue when table was not updated when groupId property changed
+### Added
+- Added new icons for users
+- Added more information for corner cases (no user)
+
 ## [0.1.0] - 2018-08-14
 ### Fixed
 - Removed modal box and replaced with inline editing